refactor: enum 생성으로 인한 코드 변경
