+++ src/main/java/common/HttpRequest.java
@@ -59,12 +59,4 @@ public class HttpRequest {
 
         return sb.toString();
     }
-
-    public enum Method {
-        GET, POST, PUT, DELETE;
-
-        public static boolean isGetMethod(Method method) {
-            return method.equals(GET);
-        }
-    }
 }
