+++ src/main/java/common/RequestLine.java
@@ -1,23 +1,21 @@
 package common;
 
-import common.HttpRequest;
-
 import java.util.Map;
 
 public class RequestLine {
-    private final HttpRequest.Method method;
+    private final Method method;
     private final String path;
     private final String version;
     private final Map<String, String> params;
 
-    public RequestLine(HttpRequest.Method method, String path, String version, Map<String, String> params) {
+    public RequestLine(Method method, String path, String version, Map<String, String> params) {
         this.method = method;
         this.path = path;
         this.version = version;
         this.params = params;
     }
 
-    public HttpRequest.Method getMethod() {
+    public Method getMethod() {
         return method;
     }
