+++ src/main/java/controller/IndexController.java
@@ -3,7 +3,7 @@ package controller;
 import modelview.ModelView;
 import common.HttpRequest;
 
-import static common.HttpRequest.Method.isGetMethod;
+import static common.Method.isGetMethod;
 
 public class IndexController implements Controller {
     @Override
