+++ src/main/java/webserver/RequestHandler.java
@@ -1,17 +1,25 @@
 package webserver;
 
+import common.HttpRequest;
+import common.HttpResponse;
+import common.Method;
 import controller.Controller;
 import modelview.ModelView;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import common.HttpRequest;
 import utils.HttpRequestUtils;
+import utils.HttpResponseUtils;
 
-import java.io.*;
+import java.io.File;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
 import java.net.Socket;
 import java.nio.file.Files;
 
-import static common.HttpRequest.Method;
+import static common.ContentType.HTML;
+import static common.ResponseCode.OK;
+
 
 public class RequestHandler implements Runnable {
     private static final Logger logger = LoggerFactory.getLogger(RequestHandler.class);
@@ -31,20 +39,23 @@ public class RequestHandler implements Runnable {
             HttpRequest request = HttpRequestUtils.createRequest(in);
             logger.debug(request.toString());
 
-            // HTTP Request를 처리할 컨트롤러 가져오기
+            // HTTP 요청을 처리할 컨트롤러 가져오기
             Controller controller = fetchController(request);
 
-            // 컨트롤러 실행 -> [모델 + 뷰] 객체 반환
+            // 컨트롤러가 요청을 처리
             ModelView modelView = controller.process(request);
 
-            // 실제 위치로 반환
+            // 파일의 논리 경로 -> 실제 경로(full-path)로 반환
             String path = PathResolver.resolve(modelView.getViewName());
+
+            // 파일 내용 불러오기
             byte[] body = Files.readAllBytes(new File(path).toPath());
 
-            // TODO : 적절한 응답 메세지 만들어서 보내기
-            // HTTP Response Message 보내기
-            sendHttpResponseMessage(out, body);
+            // Response 객체 생성
+            HttpResponse response = HttpResponseUtils.createDefaultHeaderResponse(request.getVersion(), OK, HTML, body);
 
+            // HTTP 응답 보내기
+            HttpResponseUtils.sendResponse(out, response);
         } catch (IOException e) {
             logger.error(e.getMessage());
             throw new RuntimeException(e);
@@ -70,36 +81,10 @@ public class RequestHandler implements Runnable {
         return controller;
     }
 
-    // TODO : ViewResolver와 중복되므로 하나로 빼놓기
+    // TODO : PathResolver와 중복되므로 하나로 빼놓기
     private boolean requestStaticFile(String path) {
         return path.endsWith(".css") || path.startsWith("/fonts") ||
                 path.startsWith("/images") || path.startsWith("/js") ||
                 path.endsWith(".ico");
-    }
-
-    private void sendHttpResponseMessage(OutputStream out, byte[] body) throws IOException {
-        DataOutputStream dos = new DataOutputStream(out);
-        response200Header(dos, body.length);
-        responseBody(dos, body);
-    }
-
-    private void response200Header(DataOutputStream dos, int lengthOfBodyContent) {
-        try {
-            dos.writeBytes("HTTP/1.1 200 OK \r\n");
-            dos.writeBytes("Content-Type: text/html;charset=utf-8\r\n");
-            dos.writeBytes("Content-Length: " + lengthOfBodyContent + "\r\n");
-            dos.writeBytes("\r\n");
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    private void responseBody(DataOutputStream dos, byte[] body) {
-        try {
-            dos.write(body, 0, body.length);
-            dos.flush();
-        } catch (IOException e) {
-            logger.error(e.getMessage());
-        }
     }
 }
