+++ src/main/java/webserver/WebServer.java
@@ -3,12 +3,13 @@ package webserver;
 import controller.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import common.HttpRequest.Method;
 
 import java.net.ServerSocket;
 import java.net.Socket;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
+
+import static common.Method.GET;
 
 public class WebServer {
     private static final Logger logger = LoggerFactory.getLogger(WebServer.class);
@@ -24,11 +25,11 @@ public class WebServer {
         }
 
         RequestControllerMapper mapper = RequestControllerMapper.getInstance();
-        mapper.put("/static", Method.GET, new StaticController());
-        mapper.put("/error", Method.GET, new ErrorController());
-        mapper.put("/index.html", Method.GET, new IndexController());
-        mapper.put("/user/form.html", Method.GET, new GetUserFormController());
-        mapper.put("/user/create", Method.GET, new CreateUserController());
+        mapper.put("/static", GET, new StaticController());
+        mapper.put("/error", GET, new ErrorController());
+        mapper.put("/index.html", GET, new IndexController());
+        mapper.put("/user/form.html", GET, new GetUserFormController());
+        mapper.put("/user/create", GET, new CreateUserController());
 
         // 서버소켓을 생성한다. 웹서버는 기본적으로 8080번 포트를 사용한다.
         try (ServerSocket listenSocket = new ServerSocket(port)) {
